+++ language-support/java/codegen/src/ledger-tests/scala/com/digitalasset/CodegenLedgerTest.scala
@@ -226,4 +226,25 @@ class CodegenLedgerTest extends FlatSpec with Matchers with BazelRunfiles {
     wolpertinger.key.get.owner shouldEqual "Alice"
     wolpertinger.key.get.age shouldEqual java.math.BigDecimal.valueOf(17.42)
   }
+
+  it should "be able to exercise by key" in withClient { client =>
+    sendCmd(client, glookofly.create(), sruquito.create())
+
+    // We'll exercise by key, no need to get the handles
+    val glookoflyContract :: sruquitoContract :: Nil = readActiveContracts(client)
+
+    val tob = Instant.now().`with`(ChronoField.NANO_OF_SECOND, 0)
+    val reproduceByKeyCmd =
+      Wolpertinger.exerciseByKeyReproduce(glookoflyContract.key.get, sruquitoContract.id, tob)
+    sendCmd(client, reproduceByKeyCmd)
+
+    val wolpertingers = readActiveContracts(client)
+    wolpertingers should have length 2
+
+    val sruq :: glookosruq :: Nil = wolpertingers
+
+    sruq.data.name shouldEqual sruquito.name
+    glookosruq.data.name shouldEqual s"${glookofly.name}-${sruquito.name}"
+    glookosruq.data.timeOfBirth shouldEqual tob
+  }
 }