+++ docs/source/support/release-notes.rst
@@ -21,6 +21,18 @@ Java Codegen
 - Support generic types (including tuples) as contract keys in codegen.
   See `#1728 <https://github.com/digital-asset/daml/issues/1728>`__.
 
+Ledger API
+~~~~~~~~~~
+
+- A new command ``ExerciseByKey`` allows to exercise choices on active contracts referring to them by their key.
+  See `#1366 <https://github.com/digital-asset/daml/issues/1366>`__.
+
+Java Bindings
+~~~~~~~~~~~~~
+
+- The addition of the ``ExerciseByKey`` to the Ledger API is reflected in the bindings.
+  See `#1366 <https://github.com/digital-asset/daml/issues/1366>`__.
+
 Release Procedure
 ~~~~~~~~~~~~~~~~~
 
@@ -85,9 +97,6 @@ Release Procedure
 
 - Fixes to the release procedure.
   See `#1725 <https://github.com/digital-asset/daml/issues/1725>`__
-
-Java Bindings
-~~~~~~~~~~~~~
 
 - The changes for Java Bindings listed for SDK 0.13.1 now only apply to SDK 0.13.2 and later.
   This is due to the partial failure of the release procedure.